% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTAPMain.R
\name{plotMTAPResults}
\alias{plotMTAPResults}
\title{Plot the result of MTAP .}
\usage{
plotMTAPResults(MTAPResult, neigType)
}
\arguments{
\item{MTAPResult}{result of MTAP}

\item{neigType}{queen-rook-hexagon}
}
\value{
none
}
\description{
This function plots the solution path
}
