% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTAPMain.R
\name{solveMTAPInstance}
\alias{solveMTAPInstance}
\title{Solve a MTAP .}
\usage{
solveMTAPInstance(n, neigType, upperLimit)
}
\arguments{
\item{n}{size of the surface}

\item{neigType}{queen-rook-hexagon}

\item{upperLimit}{upper limit for random cost surface}
}
\value{
MTAPResult
}
\description{
This function solves an MTAP instance
}
